scripts/requirements.txt
supabase==2.10.0
watchdog==6.0.0
python-dotenv==1.0.1
